avoid moving functions
