+++ app/client/src/actions/ideActions.ts
@@ -35,37 +35,3 @@ export const setShowQueryCreateNewModal = (payload: boolean) => {
     payload,
   };
 };
-export const closeJSActionTab = (payload: { id: string; parentId: string }) => {
-  return {
-    type: ReduxActionTypes.CLOSE_JS_ACTION_TAB,
-    payload,
-  };
-};
-export const closeJsActionTabSuccess = (payload: {
-  id: string;
-  parentId: string;
-}) => {
-  return {
-    type: ReduxActionTypes.CLOSE_JS_ACTION_TAB_SUCCESS,
-    payload,
-  };
-};
-
-export const closeQueryActionTab = (payload: {
-  id: string;
-  parentId: string;
-}) => {
-  return {
-    type: ReduxActionTypes.CLOSE_QUERY_ACTION_TAB,
-    payload,
-  };
-};
-export const closeQueryActionTabSuccess = (payload: {
-  id: string;
-  parentId: string;
-}) => {
-  return {
-    type: ReduxActionTypes.CLOSE_QUERY_ACTION_TAB_SUCCESS,
-    payload,
-  };
-};
