+++ app/client/src/actions/pluginActionActions.ts
@@ -1,10 +1,8 @@
 import type { ActionResponse, PaginationField } from "api/ActionAPI";
-import type {
-  AnyReduxAction,
-  EvaluationReduxAction,
-  ReduxAction,
-} from "@appsmith/constants/ReduxActionConstants";
 import {
+  type AnyReduxAction,
+  type EvaluationReduxAction,
+  type ReduxAction,
   ReduxActionErrorTypes,
   ReduxActionTypes,
 } from "@appsmith/constants/ReduxActionConstants";
@@ -410,4 +408,22 @@ export default {
   updateAction,
   updateActionSuccess,
   bindDataOnCanvas,
+};
+export const closeQueryActionTab = (payload: {
+  id: string;
+  parentId: string;
+}) => {
+  return {
+    type: ReduxActionTypes.CLOSE_QUERY_ACTION_TAB,
+    payload,
+  };
+};
+export const closeQueryActionTabSuccess = (payload: {
+  id: string;
+  parentId: string;
+}) => {
+  return {
+    type: ReduxActionTypes.CLOSE_QUERY_ACTION_TAB_SUCCESS,
+    payload,
+  };
 };
