+++ app/client/src/pages/Editor/IDE/hooks.ts
@@ -32,7 +32,8 @@ import { useQueryAdd } from "@appsmith/pages/Editor/IDE/EditorPane/Query/hooks";
 import { TabSelectors } from "./EditorTabs/constants";
 import { createEditorFocusInfoKey } from "@appsmith/navigation/FocusStrategy/AppIDEFocusStrategy";
 import { FocusElement } from "../../../navigation/FocusElements";
-import { closeJSActionTab, closeQueryActionTab } from "actions/ideActions";
+import { closeJSActionTab } from "../../../actions/jsActionActions";
+import { closeQueryActionTab } from "../../../actions/pluginActionActions";
 
 export const useCurrentAppState = () => {
   const [appState, setAppState] = useState(EditorState.EDITOR);
